+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Register-usage-record - distribution filter canceled subscription
+
 ## [3.0.0] - 2022-06-29
 
 ### Added